fix(API): Duplicate props are now handled.
By default we handle duplicate props by providing either the prop value from the most recently
executed Query function, otherwise if a prop was passed into the component that conflicts with a
Query function prop then the passed in prop value will be used.  So there is an order of preference
here.  An conflicResolver configuration option has been added which allows you to provide your own
logic when there are property conflicts.