perf(optimisations): ComponentQueries has been optimised to behave like a "pure" component.
BREAKING CHANGE: This is a *partially* breaking change that I don't expect will affect most users. I have extended the HOC so that it acts like a pure component.  In doing so it checks to see if either the result of the queries are new or new _additional_ props have been passed into your wrapped component. The `render` then only fires if one of these conditions were met. You can override this behaviour by using the new complex configuration option:
`componentQueries({ queries: [...], config: { pure: false }})`
The new `config` prop includes the `pure` option as shown above as well as all the properties from the `sizeMeConfig` configuration prop (which is to be deprecated).

I've had to do a placeholder commit here as I previously got the semantic-release format for commit message wrong.